Add thread safety for glob initialization.
In a parallel build environment, it is possible for multiple projects
to request the version of the same exact key. In this scenario, both
could try to initialize the glob cache at the same time, and the first
one to finish then immediately tries to iterate over the globs while
the other one is still initializing, causing a concurrent modification
exception.

Targets #27.